Add Mezzio TwigRenderer support
This change adds Mezzio TwigRenderer support to the project, which
provides a simple and (relatively) easy to use view layer for the
application. I've decided to go with Twig as it's the templating layer
that I most prefer to use. However, the project is built on
mezzio-template, so any package that builds on it can be used, if and
when desired in the future.