Add new development dependencies
This change adds three new dependencies: league/commonmark,
mikey179/vfsstream, and phpstan/phpstan. The first two are required to
convert Markdown to HTML as part of the conversion process to ODT and to
mock the filesystem for testing the created ODT file. The final one is
to add static analysis support.